Rename src/components/cylc/tree/tree.js to src/components/cylc/tree/cylc-tree.js
This matches the class inside cylc-tree.js, CylcTree. Also prevents issues
with file systems that are case insensitive (bug found by @oliver-sanders).
